Gave users authorities

+++ modules/MOD-2-BUILDING-RESTful-APIs/messages-API/src/main/java/com/mod2api/messages/apprentice/Apprentice.java
@@ -15,12 +15,12 @@ public class Apprentice {
             strategy = GenerationType.SEQUENCE,
             generator = "apprentice_sequence"
     )
-    private Long id;
+    private Integer id;
     private String name;
 
     public Apprentice(){}
 
-    public Apprentice(Long id, String name) {
+    public Apprentice(Integer id, String name) {
         this.id = id;
         this.name = name;
     }
@@ -29,11 +29,11 @@ public class Apprentice {
         this.name = name;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
